project/requirements.txt


fastapi[all]
chromadb
sentence-transformers
PyPDF2
python-docx
uvicorn
